+++ interpreter/rxvmplugin/rxvmplugins/mc_decimal/mc_decimal_full_tests.c
@@ -41,7 +41,7 @@ int aTestFromToInt(char* expected, int64_t int_input) {
         errors++;
     }
     else printf("OK - ");
-    printf("expected %lld got %lld for to int\n", int_input, int_output);
+    printf("expected %lld got %lld for to int\n", (long long)int_input, (long long)int_output);
 
     free(output);
     if (a.decimal_value) free(a.decimal_value);
@@ -171,9 +171,9 @@ static int test_int_tofrom_conversions() {
         int_input = int_input * rand(); // NOLINT(cert-msc50-cpp)
         int_input = int_input * rand(); // NOLINT(cert-msc50-cpp)
 
-        sprintf(input, "%lld", int_input);
+        sprintf(input, "%lld", (long long)int_input);
         errors += aTestFromToInt(input, int_input);
-        sprintf(input, "%lld", -int_input);
+        sprintf(input, "%lld", (long long)(-int_input));
         errors += aTestFromToInt(input, -int_input);
     }
     free(input);
@@ -417,7 +417,7 @@ int test_moreDecimalToInteger() {
         errors++;
     }
     else printf("OK - signal - ");
-    printf("decNumber: nan, Result: %lld\n", result);
+    printf("decNumber: nan, Result: %lld\n", (long long)result);
 
     // Test Infinity
     plugin->decimalFromString(plugin, &input, "inf");
@@ -427,7 +427,7 @@ int test_moreDecimalToInteger() {
         errors++;
     }
     else printf("OK - signal - ");
-    printf("decNumber: inf, Result: %lld\n", result);
+    printf("decNumber: inf, Result: %lld\n", (long long)result);
 
     // Test -Infinity
     plugin->decimalFromString(plugin, &input, "-inf");
@@ -437,7 +437,7 @@ int test_moreDecimalToInteger() {
         errors++;
     }
     else printf("OK - signal - ");
-    printf("decNumber: -inf, Result: %lld\n", result);
+    printf("decNumber: -inf, Result: %lld\n", (long long)result);
 
     // Test 0.0
     plugin->decimalFromString(plugin, &input, "0");
@@ -447,7 +447,7 @@ int test_moreDecimalToInteger() {
         errors++;
     }
     else printf("OK - ");
-    printf("decNumber: 0.0, Result: %lld\n", result);
+    printf("decNumber: 0.0, Result: %lld\n", (long long)result);
 
     // Test -0.0
     plugin->decimalFromString(plugin, &input, "-0");
@@ -457,16 +457,16 @@ int test_moreDecimalToInteger() {
         errors++;
     }
     else printf("OK - ");
-    printf("decNumber: -0.0, Result: %lld\n", result);
+    printf("decNumber: -0.0, Result: %lld\n", (long long)result);
 
     // Test 1.23456789
     plugin->decimalFromString(plugin, &input, "1.23456789");
     plugin->decimalToInt(plugin, &input, &result);
     if (!plugin->base.signal_number) { // (result != 1) {
-        printf("Error - no signal: 1.23456789, Result: %lld\n", result);
+        printf("Error - no signal: 1.23456789, Result: %lld\n", (long long)result);
         errors++;
     }
-    else printf("OK - signal: decNumber: 1.23456789, Result: %lld\n", result);
+    else printf("OK - signal: decNumber: 1.23456789, Result: %lld\n", (long long)result);
 
     // Test 1234567890123456789
     plugin->decimalFromString(plugin, &input, "1234567890123456789");
@@ -476,7 +476,7 @@ int test_moreDecimalToInteger() {
         errors++;
     }
     else printf("OK - ");
-    printf("decNumber: 1234567890123456789, Result: %lld\n", result);
+    printf("decNumber: 1234567890123456789, Result: %lld\n", (long long)result);
 
         if (input.decimal_value) free(input.decimal_value);
     return errors;
@@ -839,7 +839,7 @@ int testDecimalExtract(char* input, char* expected_coefficient, int64_t expected
         errors++;
     }
     else printf("OK - ");
-    if (exponent) printf("%s -> %se%lld\n", input, coefficient, exponent);
+    if (exponent) printf("%s -> %se%lld\n", input, coefficient, (long long)exponent);
     else printf("%s -> %s\n", input, coefficient);
 
     free(a.decimal_value);
